.github/workflows/requirements.txt
# Core LLM Streamlit Application Dependencies
streamlit
anthropic
python-dotenv